+++ src/compas_cadwork/scene/__init__.py
@@ -5,12 +5,14 @@ from compas_monosashi.sequencer import Text3d
 from compas_monosashi.sequencer import LinearDimension
 from compas_monosashi.sequencer import Model3d
 
+from .camera import Camera
 from .scene import CadworkSceneObject
 from .instructionobject import Text3dSceneObject
 from .instructionobject import LinearDimensionSceneObject
 from .instructionobject import Model3dSceneObject
 
 __all__ = [
+    "Camera",
     "CadworkSceneObject",
     "Text3dSceneObject",
     "LinearDimensionSceneObject",